Fix: compiler warnings
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8814 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Laptop/DropDown.cpp
@@ -95,8 +95,8 @@ DropDownBase::Init( UINT16 sX, UINT16 sY )
 	mFirstShownEntry = max( 0, min( mFirstShownEntry, mEntryVector.size( ) - 1 - mNumDisplayedEntries ) );
 
 	musWidth = 0;
-	UINT16 size = mEntryVector.size( );
-	for ( UINT16 i = 0; i < size; ++i )
+	size_t size = mEntryVector.size();
+	for ( size_t i = 0; i < size; ++i )
 	{
 		musWidth = max( musWidth, StringPixLength( mEntryVector[i].second, DEF_DROPDOWN_FONT ) );
 	}
